Move config structs to packages
