+++ cmd/testdata/config_invalid.toml
@@ -1,0 +1,1 @@
+foobar
