--- runtime/config/testdata/config_failed.toml
@@ -1,1 +1,0 @@
-asdas
