+++ webserver/config.go
@@ -1,0 +1,7 @@
+package webserver
+
+type Config struct {
+	Enable  bool   `toml:"enable"`
+	Bind    string `toml:"bind"`
+	Webroot string `toml:"webroot"`
+}
